siema

+++ Assets/_Project/Scripts/Player.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    public int balance = 9999999;
+    public float balance = 9999999;
     public TextMeshProUGUI balanceText;
     public List<Attraction> attractionList = new List<Attraction>();
